Extract copySourceTableToTargetTable() and add to interface
