+++ experiment/src/org/labkey/experiment/SampleTypeMigrationSchemaHandler.java
@@ -1,0 +1,78 @@
+package org.labkey.experiment;
+
+import org.apache.logging.log4j.Logger;
+import org.jetbrains.annotations.Nullable;
+import org.labkey.api.collections.CsvSet;
+import org.labkey.api.data.DatabaseMigrationService;
+import org.labkey.api.data.SQLFragment;
+import org.labkey.api.data.SimpleFilter;
+import org.labkey.api.data.SimpleFilter.FilterClause;
+import org.labkey.api.data.SimpleFilter.OrClause;
+import org.labkey.api.data.SimpleFilter.SQLClause;
+import org.labkey.api.data.TableInfo;
+import org.labkey.api.data.TableSelector;
+import org.labkey.api.exp.api.SampleTypeDomainKind;
+import org.labkey.api.query.FieldKey;
+import org.labkey.api.util.Formats;
+import org.labkey.api.util.GUID;
+import org.labkey.api.util.logging.LogHelper;
+
+import java.util.Collection;
+import java.util.Set;
+
+class SampleTypeMigrationSchemaHandler extends DatabaseMigrationService.DefaultMigrationSchemaHandler
+{
+    private static final Logger LOG = LogHelper.getLogger(SampleTypeMigrationSchemaHandler.class, "Sample type migration status");
+
+    public SampleTypeMigrationSchemaHandler()
+    {
+        super(SampleTypeDomainKind.getSchema());
+    }
+
+    @Override
+    public @Nullable FieldKey getContainerFieldKey(TableInfo table)
+    {
+        return FieldKey.fromParts("DUMMY"); // Unused dummy value -- see override below
+    }
+
+    @Override
+    public FilterClause getContainerClause(TableInfo sourceTable, FieldKey containerFieldKey, Set<GUID> containers)
+    {
+        return new SQLClause(
+            new SQLFragment("RowId IN (SELECT RowId FROM exp.Material WHERE Container")
+                .appendInClause(containers, sourceTable.getSqlDialect())
+                .append(")")
+        );
+    }
+
+    @Override
+    public void addDomainDataFilter(OrClause orClause, DatabaseMigrationService.DomainFilter filter, TableInfo sourceTable, FieldKey fKey, Set<String> selectColumnNames)
+    {
+        // Sample-type-specific optimization
+        if (filter.column().equalsIgnoreCase("Flag"))
+        {
+            // Select all rows where the built-in flag column equals the filter value
+            orClause.addClause(
+                new SQLClause(new SQLFragment(
+                    "RowId IN (SELECT RowId FROM exp.Material WHERE Container")
+                    .appendInClause(filter.containers(), sourceTable.getSqlDialect())
+                    .append(" AND ObjectId IN (SELECT ObjectId FROM exp.ObjectProperty WHERE StringValue = ? AND PropertyId = ?))")
+                    .add(filter.condition().getParamVals()[0])
+                    .add(getCommentPropertyId(sourceTable))
+                )
+            );
+        }
+        else
+        {
+            addDomainDataStandardFilter(orClause, filter, sourceTable, fKey, selectColumnNames);
+        }
+    }
+
+    @Override
+    public void afterTable(TableInfo sourceTable, TableInfo targetTable, SimpleFilter notCopiedFilter)
+    {
+        Collection<String> notCopiedLsids = new TableSelector(sourceTable, new CsvSet("LSID, RowId"), notCopiedFilter, null).getCollection(String.class);
+        if (!notCopiedLsids.isEmpty())
+            LOG.info("   {} rows not copied", Formats.commaf0.format(notCopiedLsids.size()));
+    }
+}
